requirements.txt
Django~=4.2.14
djangorestframework~=3.15.2
pdf2image~=1.17.0
pytesseract~=0.3.10
pillow~=10.4.0